# Not use predicate_refuted_by when pruning chunks