DESCRIPTION: Avoids using `predicate_refuted_by` when pruning columnar table chunks

Closes https://github.com/citusdata/citus/issues/4670.

Opening for feedback.

Since `predicate_refuted_by` is extremely costly, we want to re-use the fast pruning logic in `shard_pruning.c`.
Looked into that file and seems that we have a 4-step logic to prune shards.
To me, the only step that applies to chunk pruning is the exhaustive pruning step (step-4) due to the nature of chunks.
```c
/*-------------------------------------------------------------------------
 *
 * shard_pruning.c
 ......
 * 4) If there are overlapping shards, exhaustively search all shards that are
 *    not excluded by constraints
```

Talked with Andres, he suggested that we could first sort chunks and then could do the chunk pruning via binary search on chunk boundaries but I'm not sure if it applies to chunk pruning since I think we can't actually have a good way of sorting the chunks.
Any ideas ? (cc: @jeff-davis)

---

Also, I didn't do much benchmarking but the chunk pruning logic proposed in this pr seems to be much faster than using `predicate_refuted_by` when accessing the columnar table in a selective way:

```sql
create table local (a int) using columnar;
insert into local values (generate_series(1, 250000000));

SELECT sum(a) FROM local WHERE a = 10000325 or a = 104356 or a = 765567;
-- on this branch: ~95ms
-- on master: ~400ms
```

Is there a better way to measure the improvement done in this pr ?

-----

First two commits are to partially re-use the pruning logic from `shard_pruning.c`.
In the second commit, note that I removed two shard pruning specific fields from `ClauseWalkerContext` (`partitionMethod` & `compareIntervalFunctionCall`). This is because:
a) To re-use the same logic in chunk pruning when generating prune instances
b) Those two fields anyway are not used when generating prune instances for shard pruning and the only place that makes sense to use `ClauseWalkerContext` was already `PrunableExpressions`

Summarized design considerations & TODO's:
- [ ] Look more into `postgres/partprune.c` to see if we could directly use postgres's partitioning logic
- [ ] Add regression tests to cover each case in `ExhaustivePruneOneWithMinMax` by columnar tables
- [ ] Do some reasonable benchmarking
- [ ] Another improvement: prune chunk groups at once (https://github.com/citusdata/citus/pull/4753#discussion_r584541765) (cc: @pykello)
- [ ] Could we have have a more advanced pruning logic in chunk pruning too ?
This pr only uses exhaustive pruning (step-4) as mentioned above.