@jeff-davis 2021-03-01
The chunks can be sorted by min/max, right? Why can't we have a good way to sort them?

Also, I'd still like to keep in mind the goals of this PR. Is it simplicity, performance, or both? If it adds to complexity, I think we should be cautious about proceeding and we definitely need some numbers to back it up.