@onurctirtir 2021-03-02
> The chunks can be sorted by min/max, right? Why can't we have a good way to sort them?

True, we could first sort chunks and then could use `PruneWithBoundaries` etc. .. before exhaustive pruning.

>  Is it simplicity, performance, or both? If it adds to complexity, I think we should be cautious about proceeding

That step-4 was maybe the simplest part to implement, but to be able to apply the other steps in shard pruning (like first sorting then calling `PruneWithBoundaries`), I think we first would have to do some more refactoring but then it would add an important amount of complexity to the pr.

What do you think @jeff-davis ? Should we skip this pr for now ?