@onurctirtir 2021-03-03
>Also, I didn't do much benchmarking but the chunk pruning logic proposed in this pr seems to be much faster than using >predicate_refuted_by when accessing the columnar table in a selective way:
>
>create table local (a int) using columnar;
>insert into local values (generate_series(1, 250000000));
>
>SELECT sum(a) FROM local WHERE a = 10000325 or a = 104356 or a = 765567;
>-- on this branch: ~95ms
>-- on master: ~400ms


In addition to above measurement::

-- prune all chunks
SELECT * FROM local WHERE a > 400000000;
-- on this branch: ~80ms
-- on master: ~200ms

-- count %5 of whole data
SELECT COUNT(*) FROM local WHERE a < 12500000;
-- might be noise
-- on this branch: ~2500ms
-- on master: ~2600ms

Maybe after applying other steps from shard pruning (i.e optimized ones, like doing binary search over sorted chunk boundaries), we could improve performance gain even more for _selective_ queries. But for now, we decided to defer this pr to another time since it would require an important amount of refactor in `shard_pruning.c`